Game files not found error handling

+++ NoitaSaves.py
@@ -33,38 +33,43 @@ while scenario != 'e':
         else:
             print('Error: Incorrect scenario')
 
-    if scenario == 's':
-        save_name_correct = False
-        while not save_name_correct:
-            save_name = input('Input save name >> ')
-            save_name_errors = set(re.findall(r'[^A-Za-zА-Яа-я0-9\- ]', save_name))
-            if len(save_name_errors) > 0:
-                print('Error: Incorrect symbols: [', end='')
-                print(*save_name_errors, sep='], [', end=']\n')
-            else:
-                save_name = save_name.replace(' ', '_')
-                if save_name in os.listdir(saves_dir):
-                    print('Error: This save is already exists')
-                else:
-                    save_name_correct = True
-        shutil.copytree(game_dir + r'\save00', saves_dir + '\\' + save_name)
-
-    if scenario in ('l', 'd'):
-        save_index_correct = False
-        while not save_index_correct:
-            save_index = int(input('Select the save index >> '))
-            if 0 < save_index <= len(saves):
-                save_index_correct = True
-            else:
-                print('Error: Incorrect index')
-        if scenario == 'l':
-            shutil.rmtree(game_dir + r'\save00')
-            shutil.copytree(saves_dir + '\\' + saves[save_index - 1], game_dir + r'\save00')
-        if scenario == 'd':
-            shutil.rmtree(saves_dir + '\\' + saves[save_index - 1])
-
     if scenario in ('s', 'l', 'd'):
-        print('\nDone!\n\n')
+
+        if not os.path.exists(game_dir):
+            print('\nError: Game files not found\n\n')
+        else:
+
+            if scenario == 's':
+                save_name_correct = False
+                while not save_name_correct:
+                    save_name = input('Input save name >> ')
+                    save_name_errors = set(re.findall(r'[^A-Za-zА-Яа-я0-9\- ]', save_name))
+                    if len(save_name_errors) > 0:
+                        print('Error: Incorrect symbols: [', end='')
+                        print(*save_name_errors, sep='], [', end=']\n')
+                    else:
+                        save_name = save_name.replace(' ', '_')
+                        if save_name in os.listdir(saves_dir):
+                            print('Error: This save is already exists')
+                        else:
+                            save_name_correct = True
+                shutil.copytree(game_dir + r'\save00', saves_dir + '\\' + save_name)
+
+            elif scenario in ('l', 'd'):
+                save_index_correct = False
+                while not save_index_correct:
+                    save_index = int(input('Select the save index >> '))
+                    if 0 < save_index <= len(saves):
+                        save_index_correct = True
+                    else:
+                        print('Error: Incorrect index')
+                if scenario == 'l':
+                    shutil.rmtree(game_dir + r'\save00')
+                    shutil.copytree(saves_dir + '\\' + saves[save_index - 1], game_dir + r'\save00')
+                if scenario == 'd':
+                    shutil.rmtree(saves_dir + '\\' + saves[save_index - 1])
+
+            print('\nDone!\n\n')
 
     if scenario in ('cs-d', 'cs-w', 'rs-d', 'rs-w'):
         if '-d' in scenario:
